+++ Beta Version/main.cpp
@@ -1,9 +1,29 @@
 #include "winsimple.h"
+#include <fstream>
 
 
+std::vector<char> loadIntoMemory(std::string file_path)
+{
+    std::ifstream file(file_path, std::ios::binary | std::ios::ate); // Open at end to get size
 
+    if (!file.is_open()) {
+        std::cerr << "Failed to open file: " << file_path << std::endl;
+        std::vector<char> empty;
+        return empty;
+    }
 
+    // Get file size
+    std::streamsize size = file.tellg();
+    file.seekg(0, std::ios::beg); // Return to start
 
+    std::vector<char> buffer(size);
+
+    file.read(buffer.data(), size);
+
+    file.close();
+    
+    return buffer;
+}
 
 int main()
 {
@@ -11,16 +31,36 @@ int main()
 	ws::Window window;
 	
 	
-	ws::Texture backTex;
-	ws::Sprite sprite;
-	backTex.load("back2.jpg");
+	ws::Texture backTex,backTex2;
+	ws::Sprite sprite,sprite2;
+	
+	backTex.loadFromFile("back2.jpg");
+
+	std::vector<char> data = loadIntoMemory("back.bmp");
+	
+	backTex2.loadFromMemory(data.data(),data.size());
+	
+	
 	sprite.setTexture(backTex);
-	sprite.setOrigin(0,sprite.getTextureRect().height);
-	sprite.setScale(1,-1);
+	sprite2.setTexture(backTex2);
+	sprite2.y = 300;
 	
+	window.create(backTex.getSize().x,backTex.getSize().y,"");
+	//window.setFullscreen(true);
 	
-	window.create(backTex.width,backTex.height,"");
-	window.setFullscreen(true);
+	ws::View view,view2;
+	
+	view = window.getView();
+	
+	view2.setRect({0,0,backTex.getSize().x,backTex.getSize().y});
+	view2.setPortRect(0,0,backTex.getSize().x/2,backTex.getSize().y/2);
+
+
+	for(int a=0;a<backTex2.getSize().x * backTex2.getSize().y;a++)
+	{
+		backTex2.setPixel(a,Gdiplus::Color(255,rand()%255,rand()%255,rand()%255));
+	}
+	
 	
 	while(window.isOpen())
 	{
@@ -39,10 +79,18 @@ int main()
 			
 		}
 		
+		view2.setZoom(view2.getZoom() + 0.03);
 		
+		
+
 		
 		window.clear();
+		window.setView(view);
 		window.draw(sprite);
+		window.draw(sprite2);
+		window.setView(view2);
+		window.draw(sprite);
+		window.draw(sprite2);
 		window.display();
 	}